set room entered booleans and load accordingly
