+++ bathroom.lua
@@ -1,1 +1,6 @@
 --Holds bathroom state, 2 walls(shower and sink walls?)
+
+Gfx = love.graphics
+Bathroom = {}
+--by default, you have not "entered" the room. This must be set to true by walking through one of the doors to the room.
+Bathroom.entered = false
