+++ bedroom.lua
@@ -1,1 +1,6 @@
 --5  views (towards living room door, towards bathroom, towards bed, towards closet, towards desk), state of closet door, desk, 
+
+Gfx = love.graphics
+Bedroom = {}
+--by default, you have not "entered" the room. This must be set to true by walking through one of the doors to the room.
+Bedroom.entered = false
