+++ closetOne.lua
@@ -1,1 +1,6 @@
 --closet one (bedroom closet)
+
+Gfx = love.graphics
+ClosetOne = {}
+--by default, you have not "entered" the room. This must be set to true by walking through one of the doors to the room.
+ClosetOne.entered = false
