+++ closetTwo.lua
@@ -1,1 +1,6 @@
 --closet two (living room closet)
+
+Gfx = love.graphics
+ClosetTwo = {}
+--by default, you have not "entered" the room. This must be set to true by walking through one of the doors to the room.
+ClosetTwo.entered = false
