+++ kitchen.lua
@@ -1,1 +1,6 @@
 --holds kitchen state, kitchen light, laundry room locked, cabinets, dishwasher, blinds/window
+
+Gfx = love.graphics
+Kitchen = {}
+--by default, you have not "entered" the room. This must be set to true by walking through one of the doors to the room.
+Kitchen.entered = false
