+++ laundry.lua
@@ -1,1 +1,6 @@
 --state of laundry room, if door is unlocked, washer/dryer
+
+Gfx = love.graphics
+Laundry = {}
+--by default, you have not "entered" the room. This must be set to true by walking through one of the doors to the room.
+Laundry.entered = false
